move functions to test utils
