+++ corehq/ex-submodules/casexml/apps/phone/tests/test_state_hash.py
@@ -4,12 +4,11 @@ from casexml.apps.case.exceptions import BadStateException
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.phone.models import SyncLog, User
 from datetime import datetime
-from casexml.apps.phone.restore import generate_restore_payload,\
-    generate_restore_response
 from casexml.apps.phone.checksum import EMPTY_HASH, CaseStateHash
 from casexml.apps.case.xml import V2
 from casexml.apps.case.util import post_case_blocks
 from casexml.apps.case.tests.util import delete_all_sync_logs, delete_all_xforms, delete_all_cases
+from casexml.apps.phone.tests.utils import generate_restore_payload, generate_restore_response
 from corehq import toggles
 from toggle.shortcuts import update_toggle_cache, clear_toggle_cache
